All-round improvements
* MANY improvements to logic.

* Cosmetic improvements to code style, in particular indentation depth,
  which should really be kept under 80 cols.

  (no changes made to tab offset since 4 seems to be author's personal
  preference)

* Directive functor declared separately for increased readability.

* Breaking change: renamed `abstractProxyProperty´ to
  `abstractproxyProperty´ for consistency with `displaynameProperty´
  attribute.

--

NOTE

This commit is a *manual integration* of modifications I made to
the original uiBreadcrums code, which I felt I should contribute back
for obvious reasons.  My modifications work flawlessly but I have not
had the opportunity to ensure that *this integration* actually works.
Please test.

Whether this commit is accepted or not, many thanks to the author for
this solution on which this Angular-noob was able to quickly craft a
working breadcrumbs implementation.

 - Miguel